import fixes

+++ app/Console/Commands/ImportProductsCommand.php
@@ -63,7 +63,7 @@ class ImportProductsCommand extends Command
             'name' => $product->title,
             'brand_id' => Brand::whereNodId($product->manufacturer_id)->first()->id ?? null,
             'warranty' => $product->warranty,
-            'price' => $product->promo_price,
+            'price' => $product->ron_promo_price,
             'short_description' => $product->description,
             'description' => $product->long_description,
             'sku' => $product->code,
@@ -72,9 +72,13 @@ class ImportProductsCommand extends Command
 
         $productCategoryId = $product->product_category_id;
         $newProduct = Product::updateOrCreate(['nod_id' => $product->id], $values);
-        $categoryIds = Category::getNestCategoryBy($productCategoryId, [$productCategoryId]);
 
-        $newProduct->categories()->sync($categoryIds);
+        if(! $newProduct->categories->pluck('id')->contains($productCategoryId)) {
+            $categoryIds = Category::getNestCategoryBy($productCategoryId, [$productCategoryId]);
+    
+            $newProduct->categories()->sync($categoryIds);
+        }
+        
         $this->handleImages($product->pictures, $newProduct);
     }
 
@@ -85,22 +89,25 @@ class ImportProductsCommand extends Command
             $location = $key == 0 ? 'base_image' : 'additional_images';
             $url = $image->url_overlay_picture;
             $name = substr($url, strrpos($url, '/') + 1);
-            $path = "media/{$location}/{$name}";
-            Storage::disk($disk)->put($path, file_get_contents($url));
-            $file = new SymfonyFile(public_path("storage/{$path}"));
 
-            $newFile = File::create([
-                'user_id' => 1,
-                'disk' => 'public_storage',
-                'location' => $location,
-                'filename' => $name,
-                'path' => $path,
-                'extension' => $file->guessExtension() ?? '',
-                'mime' => $file->getMimeType(),
-                'size' => $file->getSize(),
-            ]);
+            if(! $product->files()->whereFilename($name)->exists()) {
+                $path = "media/{$location}/{$name}";
+                Storage::disk($disk)->put($path, file_get_contents($url));
+                $file = new SymfonyFile(public_path("storage/{$path}"));
 
-            $product->files()->attach($newFile, ['zone' => $location, 'entity_type' => get_class($product)]);
+                $newFile = File::create([
+                    'user_id' => 1,
+                    'disk' => 'public_storage',
+                    'location' => $location,
+                    'filename' => $name,
+                    'path' => $path,
+                    'extension' => $file->guessExtension() ?? '',
+                    'mime' => $file->getMimeType(),
+                    'size' => $file->getSize(),
+                ]);
+
+                $product->files()->attach($newFile, ['zone' => $location, 'entity_type' => get_class($product)]);
+            }
         }
     }
 }
